refactor(fleet_monitoring): extract bus/api.py and keep __init__.py as package facade
Move socket I/O functions (publish, subscribe) and DEFAULT_BUS_SOCKET_PATH
from bus/__init__.py to bus/api.py per AGENTS.md package facade rules,
re-exporting public symbols from __init__.py.
